+++ apps/native-component-list/src/screens/UI/FormScreen.ios.tsx
@@ -72,7 +72,7 @@ export default function FormScreen() {
             label="Select a color"
             selection={color}
             supportsOpacity
-            onValueChanged={setColor}
+            onSelectionChange={setColor}
           />
           <Picker
             label="Menu picker"
@@ -111,7 +111,7 @@ export default function FormScreen() {
             label="Clear Image Cache"
           />
           <DisclosureGroup
-            onStateChange={setDisclosureGroupExpanded}
+            onIsExpandedChange={setDisclosureGroupExpanded}
             isExpanded={disclosureGroupExpanded}
             label="Show User Profile Details">
             <Text>Name: John Doe</Text>