+++ apps/native-component-list/src/screens/UI/UIScreen.ios.tsx
@@ -115,6 +115,14 @@ export const UIScreens = [
     },
   },
   {
+    name: 'Section component',
+    route: 'ui/section',
+    options: {},
+    getComponent() {
+      return optionalRequire(() => require('./SectionScreen'));
+    },
+  },
+  {
     name: 'BottomSheet component',
     route: 'ui/bottomsheet',
     options: {},